fix(initial-unsplash-integration): moved helpers out of page folder
