+++ pages/api/get-topic-photos.js
@@ -1,5 +1,5 @@
 import attempt from "@assertchris/attempt-promise";
-import { unsplash } from "../lib";
+import { unsplash } from "../../helpers";
 
 const GetTopicPhotos = async (req, res) => {
   const [error, response] = await attempt(
